Addressed review comments

+++ api/v1alpha1/hardwareclassification_types.go
@@ -115,12 +115,12 @@ type Ram struct {
 	// +kubebuilder:validation:Minimum=1
 	// MinimumSizeGB of Ram should be greater than 0
 	// Ex. MinimumSizeGB > 0
-	MinimumSizeGB int64 `json:"minimumSizeGB,omitempty"`
+	MinimumSizeGB int `json:"minimumSizeGB,omitempty"`
 	// +optional
 	// +kubebuilder:validation:Minimum=1
 	// MaximumSizeGB should be greater than 0 or greater than MinimumSizeGB
 	// Ex. MaximumSizeGB > 0 && MaximumSizeGB > MinimumSizeGB
-	MaximumSizeGB int64 `json:"maximumSizeGB,omitempty"`
+	MaximumSizeGB int `json:"maximumSizeGB,omitempty"`
 }
 
 // ProfileMatchStatus represents the state of the HardwareClassification
@@ -129,11 +129,9 @@ type ProfileMatchStatus string
 const (
 	// ProfileMatchStatusEmpty is the default status value
 	ProfileMatchStatusEmpty ProfileMatchStatus = ""
-
 	// ProfileMatchStatusMatched is the status value when the profile
 	// matches to one of the BareMetalHost.
 	ProfileMatchStatusMatched ProfileMatchStatus = "matched"
-
 	// ProfileMatchStatusUnMatched is the status value when the profile
 	// does not match to one of the BareMetalHost.
 	ProfileMatchStatusUnMatched ProfileMatchStatus = "unmatched"
@@ -147,19 +145,15 @@ const (
 	// LabelUpdateFailure is an error condition occurring when the
 	// controller is unable to update label of BareMetalHost.
 	LabelUpdateFailure ErrorType = "label update error"
-
 	// LabelDeleteFailure is an error condition occurring when the
 	// controller is unable to delete label of BareMetalHost.
 	LabelDeleteFailure ErrorType = "label delete error"
-
 	// FetchBMHListFailure is an error condition occurring when the
 	// controller is unable to fetch BMH from BMO
 	FetchBMHListFailure ErrorType = "fetch BMH error"
-
 	// ProfileMisConfigured is an error condition occurring when the
 	// extracted profile is empty.
 	ProfileMisConfigured ErrorType = "Empty Profile Error"
-
 	// Empty is an empty error
 	Empty ErrorType = ""
 )
@@ -176,10 +170,8 @@ type HardwareClassificationStatus struct {
 
 	// ErrorType indicates the type of failure encountered
 	ErrorType ErrorType `json:"errorType,omitempty"`
-
 	// ProfileMatchStatus identifies whether a applied profile is matches or not
 	ProfileMatchStatus ProfileMatchStatus `json:"profileMatchStatus,omitempty"`
-
 	// The last error message reported by the hardwareclassification system
 	ErrorMessage string `json:"errorMessage,omitempty"`
 }
